moved implementation of web handler
